tests/wpt/web-platform-tests/css/css-anchor-position/anchor-name-002.html

<!DOCTYPE html>
<link rel="help" href="https://drafts.csswg.org/css-anchor-1/#determining">
<link rel="author" href="mailto:kojii@chromium.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/check-layout-th.js"></script>
<script src="support/test-common.js"></script>
<style>
.relpos {
  position: relative;
}
.abspos {
  position: absolute;
}
.anchor1 {
  anchor-name: --a1;
  width: 10px;
  height: 10px;
  background: orange;
}
.target {
  position: absolute;
  width: anchor-size(--a1 width);
  height: 10px;
  background: lime;
}
</style>
<body onload="checkLayoutForAnchorPos('.target')">
  <div class="relpos">
    <div>
      <div class="relpos">
        <!-- This target should not find the anchor, because the last containing
          block has `position: absolute` and is after in tree order. -->
        <div class="target" data-expected-width=0></div>
        <div class="abspos">
          <div class="relpos">
            <!-- This target should not find the anchor, because the anchor is
              absolutely positioned after it. -->
            <div class="target" data-expected-width=0></div>
            <div class="anchor1" style="position: absolute"></div>
            <!-- This target should find the anchor, because the anchor is
              absolutely positioned before it. -->
            <div class="target" data-expected-width=10></div>
          </div>
          <!-- This target should find the anchor, because the last containing
            block has `position: relative`. -->
          <div class="target" data-expected-width=10></div>
        </div>
        <!-- This target should find the anchor, because the last containing
          block has `position: absolute` and is before in tree order. -->
        <div class="target" data-expected-width=10></div>
      </div>
    </div>
    <!-- This target should find the anchor, because the last containing block
      is statically positioned. -->
    <div class="target" data-expected-width=10></div>
  </div>
</body>
